Encoder Code for the Arm Motor
The inputs from the triggers multiplied by 100 as there are 1440 ticks per revolution for the arm motor.

Added the line where the arm motor is set to mode RUN_USING_ENCODER.

Moved the code that sets the target and runs to that position to the method that would be in the loop method.